Refactor yaml parser. Separate manifest and deployment parser to different go module To make it more clear.
This is about issue #64.
