+++ utils/yamlparser.go
@@ -2,9 +2,6 @@ package utils
 
 import (
 	"github.com/openwhisk/openwhisk-client-go/whisk"
-	"gopkg.in/yaml.v2"
-	"log"
-	"strings"
 )
 
 // structs that denotes the sample manifest.yaml, wrapped yaml.v2
@@ -12,7 +9,7 @@ func NewYAMLParser() *YAMLParser {
 	return &YAMLParser{}
 }
 
-type ParseDeployerYaml interface {
+type ParseYaml interface {
 	Unmarshal(input []byte, deploy *ManifestYAML) error
 	Marshal(manifest *ManifestYAML) (output []byte, err error)
 
@@ -114,16 +111,6 @@ type ManifestYAML struct {
 	Filepath string  //file path of the yaml file
 }
 
-//********************Action functions*************************//
-func (action *Action) ComposeWskAction(manipath string) (*whisk.Action, error) {
-	wskaction, err := CreateActionFromFile(manipath, action.Location)
-	Check(err)
-	wskaction.Name = action.Name
-	wskaction.Version = action.Version
-	wskaction.Namespace = action.Namespace
-	return wskaction, err
-}
-
 //********************Trigger functions*************************//
 //add the key/value array as the annotations of the trigger.
 func (trigger *Trigger) ComposeWskTrigger(kvarr []whisk.KeyValue) *whisk.Trigger {
@@ -191,138 +178,4 @@ func (pkg *Package) GetFeedList() []Feed {
 		s1 = append(s1, feed)
 	}
 	return s1
-}
-
-//********************Application functions*************************//
-//This is for parse the deployment yaml file.
-func (app *Application) GetPackageList() []Package {
-	var s1 []Package = make([]Package, 0)
-	for pkg_name, pkg := range app.Packages {
-		pkg.Packagename = pkg_name
-		s1 = append(s1, pkg)
-	}
-	return s1
-}
-
-//********************YAMLParser functions*************************//
-func (dm *YAMLParser) ParseManifest(mani string) *ManifestYAML {
-	mm := NewYAMLParser()
-	maniyaml := ManifestYAML{}
-	content, err := new(ContentReader).LocalReader.ReadLocal(mani)
-	Check(err)
-	err = mm.Unmarshal(content, &maniyaml)
-	maniyaml.Filepath = mani
-	return &maniyaml
-}
-
-func (dm *YAMLParser) ParseDeployment(dply string) *DeploymentYAML {
-	dplyyaml := DeploymentYAML{}
-	content, err := new(ContentReader).LocalReader.ReadLocal(dply)
-	Check(err)
-	err = dm.UnmarshalDeployment(content, &dplyyaml)
-	dplyyaml.Filepath = dply
-	return &dplyyaml
-}
-
-func (dm *YAMLParser) ComposeActions(manipath string) ([]ActionRecord, error) {
-	mani := dm.ParseManifest(manipath)
-	var s1 []ActionRecord = make([]ActionRecord, 0)
-	for _, action := range mani.Package.Actions {
-		wskaction, err := CreateActionFromFile(manipath, action.Location)
-		Check(err)
-
-		record := ActionRecord{wskaction, action.Location}
-		s1 = append(s1, record)
-	}
-	return s1, nil
-}
-
-func (dm *YAMLParser) ComposeTriggers(mani string, deploy string) ([]*whisk.Trigger, error) {
-	mm := NewYAMLParser()
-	manifest := mm.ParseManifest(mani)
-	//parse the deployment yaml to get feed info as annotations for triggers
-	deployment := mm.ParseDeployment(deploy)
-	//This is to get the pkgs in deployment yaml
-	var kva []whisk.KeyValue = make([]whisk.KeyValue, 0)
-	for _, dep_pkg := range deployment.Application.GetPackageList() {
-		for _, feed := range dep_pkg.GetFeedList() {
-			log.Printf("feed name is %v", feed.Name)
-			feedFullName := strings.Join([]string{dep_pkg.Packagename, feed.Name}, "/")
-			kv := whisk.KeyValue{"feed", feedFullName}
-			kva = append(kva, kv)
-			for k, v := range feed.Inputs {
-				kv := whisk.KeyValue{k, v}
-				kva = append(kva, kv)
-			}
-		}
-	}
-
-	//parse the manifest yaml
-	var t1 []*whisk.Trigger = make([]*whisk.Trigger, 0)
-	pkg := manifest.Package
-	for _, trigger := range pkg.GetTriggerList() {
-		wsktrigger := trigger.ComposeWskTrigger(kva)
-		t1 = append(t1, wsktrigger)
-	}
-	return t1, nil
-}
-
-func (dm *YAMLParser) ComposeRules(mani string) ([]*whisk.Rule, error) {
-	mm := NewYAMLParser()
-	manifest := mm.ParseManifest(mani)
-	var r1 []*whisk.Rule = make([]*whisk.Rule, 0)
-	pkg := manifest.Package
-	for _, rule := range pkg.GetRuleList() {
-		wskrule := rule.ComposeWskRule()
-		r1 = append(r1, wskrule)
-	}
-
-	return r1, nil
-}
-
-// Is we consider multi pacakge in one yaml?
-func (dm *YAMLParser) ComposePackage(manipath string) (*whisk.SentPackageNoPublish, error) {
-	mani := dm.ParseManifest(manipath)
-	pag := &whisk.SentPackageNoPublish{}
-	pag.Name = mani.Package.Packagename
-	//The namespace for this package is absent, so we use default guest here.
-	pag.Namespace = mani.Package.Namespace
-	pag.Publish = false
-	return pag, nil
-}
-
-func (dm *YAMLParser) Unmarshal(input []byte, manifest *ManifestYAML) error {
-	err := yaml.Unmarshal(input, manifest)
-	if err != nil {
-		log.Fatalf("error happened during unmarshal :%v", err)
-		return err
-	}
-	return nil
-}
-
-func (dm *YAMLParser) UnmarshalDeployment(input []byte, deploy *DeploymentYAML) error {
-	err := yaml.Unmarshal(input, deploy)
-	if err != nil {
-		log.Fatalf("error happened during unmarshal :%v", err)
-		return err
-	}
-	return nil
-}
-
-func (dm *YAMLParser) Marshal(manifest *ManifestYAML) (output []byte, err error) {
-	data, err := yaml.Marshal(manifest)
-	if err != nil {
-		log.Fatalf("err happened during marshal :%v", err)
-		return nil, err
-	}
-	return data, nil
-}
-
-func (dm *YAMLParser) MarshalDeployment(deployment *DeploymentYAML) (output []byte, err error) {
-	data, err := yaml.Marshal(deployment)
-	if err != nil {
-		log.Fatalf("err happened during marshal :%v", err)
-		return nil, err
-	}
-	return data, nil
 }
